PWM works! (To make rob happy) Need to add bound check

+++ firmware/PWM Controller/hardware.h
@@ -35,9 +35,9 @@ this sets the port directions. Port1.6 and 1.7 are set to input because they wil
 Defining constants for the PWM board.
 **/
 #define PERIOD 60000
-#define TICKS_PER_MS (int)(PERIOD/20)
-#define MIN_PULSE (int)(0.8*TICKS_PER_MS)
-#define MIDDLE_PULSE (int)(1.5*TICKS_PER_MS)
+#define TICKS_PER_MS (uint16_t)(PERIOD/20)
+#define MIN_PULSE (uint16_t)(0.8*TICKS_PER_MS)
+#define MIDDLE_PULSE (uint16_t)(1.5*TICKS_PER_MS)
 #define SERVO_NUMBER 6
 
 void initialiseServoArray(unsigned int pulse_width);
